07_PERSONAL_DATA/requirements.txt
# Core AI and ML libraries
openai>=1.0.0
anthropic>=0.7.0
transformers>=4.30.0
torch>=2.0.0
tokenizers>=0.13.0

# Web interfaces and deployment
streamlit>=1.28.0
gradio>=3.40.0
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0

# AI frameworks and tools
langchain>=0.0.300
langchain-openai>=0.0.5
langchain-community>=0.0.10
huggingface-hub>=0.16.0

# Data handling and utilities
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0

# Optional: Advanced features
pinecone-client>=2.2.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
datasets>=2.14.0

# Development and testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0

# Discord bot (optional)
discord.py>=2.3.0

# Voice capabilities (optional)
speechrecognition>=3.10.0
pyttsx3>=2.90

# Image processing (optional)
pillow>=10.0.0
opencv-python>=4.8.0

# Jupyter notebooks (optional)
jupyter>=1.0.0
ipykernel>=6.25.0
